pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "isodate>=0.7.2",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "tqdm>=4.67.1",
]
